requirements.txt
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
PyYAML==6.0.2
random-password-generator==2.2.0
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
Werkzeug==3.1.3
